Break up the googlefonts specification
This extracts generally useful tests into their own files named after
the tables they primarily test. As these tests are imported back in,
nothing changes for the user.

Adjust tests.
